Moved classes to SiteBuilderAPI
