--- layers/API/packages/migrate-api/migrate/library/functions/load.php
@@ -1,3 +1,0 @@
-<?php
-
-require_once 'api-utils.php';
