+++ layers/API/packages/migrate-api/migrate/library/load.php
@@ -1,3 +1,2 @@
 <?php
 
-require_once 'functions/load.php';
